Revert "Swipeable routes"
